% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_rev_gwr.R
\name{rev_gwr}
\alias{rev_gwr}
\title{rev_gwr}
\usage{
rev_gwr(egid)
}
\arguments{
\item{egid:}{the egid number of the building}

\item{street}{the street name of the building (not yet available)}

\item{street_nr}{the street nr of the building (not yet available)}

\item{plz}{the postal code (not yet available)}
}
\value{
a tibble
}
\description{
The ouput of the function is a tibble with the buildings caracterictics from housing-stat.ch. Either use egid number or the postal address.
}
