% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_sun_day.R
\name{rev_sun_day}
\alias{rev_sun_day}
\title{rev_sun_day}
\usage{
rev_sun_day(location, day_of_year, ray_length, ray_intervals, zoom_level)
}
\arguments{
\item{day_of_year}{is the # of day in the year, 1 = 1. January,}

\item{ray_length}{is radius of interest, arbitrary 10km (m)}

\item{ray_intervals}{is the distance between point on one ray to check, arbitrary 100m (m)}

\item{zoom_level}{is the accuracy of the raster to get elevation from a point. 15 is the highest but it is very slow.}

\item{location:}{need to be a vector = c(long, lat) according to WGS 84,}
}
\value{
tbl_sun_shadow is a tibble
}
\description{
The ouput of the function is a tibble given the suns every minute position and if there is direct sunlight (obstructed by mountains/hills). Currently does not work on arm because of elevatr which uses sf.
}
