% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_map_lv03_wgs_lat.R
\name{rob_map_lv03_wgs_lat}
\alias{rob_map_lv03_wgs_lat}
\title{rob_map_lv03_wgs_lat}
\usage{
rob_map_lv03_wgs_lat(y, x)
}
\arguments{
\item{y:}{input coordinate y of lv03}

\item{x:}{input coordinate x of lv03}
}
\value{
latitude value of wgs84 coordinate
}
\description{
rob_map_lv03_wgs_lat computes the latitude value of wgs84
Use in combination with rob_map_lv03_wgs_lon
}
\details{
For converting lv03 to wgs84 and vice-versa
}
