% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_map_lv03_wgs_lon.R
\name{rob_map_lv03_wgs_lon}
\alias{rob_map_lv03_wgs_lon}
\title{rob_map_lv03_wgs_lon}
\usage{
rob_map_lv03_wgs_lon(y, x)
}
\arguments{
\item{y:}{input coordinate y of lv03}

\item{x:}{input coordinate x of lv03}
}
\value{
longitude value of wgs84 coordinate
}
\description{
rob_map_lv03_wgs_lon computes the longitude value of wgs84
Use in combination with rob_map_lv03_wgs_lat
}
\details{
For converting lv03 to wgs84 and vice-versa
}
