% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_map_wgs_lv03_x.R
\name{rob_map_wgs_lv03_x}
\alias{rob_map_wgs_lv03_x}
\title{rob_map_wgs_lv03_x}
\usage{
rob_map_wgs_lv03_x(lat, lon)
}
\arguments{
\item{lat:}{input coordinate latitude}

\item{lon:}{input coordinate longitude}
}
\value{
x value of lv03 coordinate
}
\description{
rob_map_wgs_lv03_x computes the x value of lv03
Use in combination with rob_map_wgs_lv03_y
}
\details{
For converting lv03 to wgs84 and vice-versa
Original: https://github.com/ValentinMinder/Swisstopo-WGS84-LV03/blob/master/scripts/r/WGS84_CH1903.R
}
