% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_str_reverse.R
\name{rob_str_reverse}
\alias{rob_str_reverse}
\title{str_reverse}
\usage{
rob_str_reverse(x)
}
\arguments{
\item{x:}{input string}
}
\value{
string with its component in reversed order
}
\description{
str_reverse reverses the order of a strings component
}
\examples{
rob_str_reverse("abc")
"cba"
}
